fix pandas _to_python_type np.float64 typo
The second branch of `_to_python_type` was checking np.int64 twice —
a copy-paste from the first branch — making the float-unwrap path
dead. Consumers (Series.apply, row iteration) were handed np.float64
instead of Python float.

Adds test coverage exercising the previously-dead branch via
Series([float,...]).apply.

+++ hipy/lib/pandas/__init__.py
@@ -48,7 +48,7 @@ def _to_native_type(x):
 def _to_python_type(x):
     if intrinsics.isa(x, np.int64):
         return int(x)
-    elif intrinsics.isa(x, np.int64):
+    elif intrinsics.isa(x, np.float64):
         return float(x)
     else:
         return x
